Refactor render node output consumer test code
Refactor TestRenderNodeOutputConsumer into its own file so that it can
be used in other tests.
